pyproject.toml
[project]
name = "template-streamlit"
version = "0.0.1"
description = "A GitHub template repository for Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.25.1",
    "httpx>=0.28.1",
    "langchain-community>=0.3.25",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.18",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.4.8",
    "msgraph-sdk>=1.31.0",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "typer>=0.16.0",
    "youtube-transcript-api>=1.0.3",
]

[project.optional-dependencies]
docs = [
    "mkdocs-material>=9.6.12",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.8.0",
    "jupyterlab>=4.4.2",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.7",
    "ty>=0.0.1a6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["D203"]

[tool.pytest.ini_options]
addopts = "-ra --cov"
testpaths = ["tests"]
pythonpath = ['.']
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.report]
show_missing = true

[tool.ty]
environment = { python-version = "3.10" }

[tool.ty.rules]
unknown-argument = "ignore" # Ignore unknown arguments in function calls
no-matching-overload = "ignore" # Ignore no matching overloads for function calls
unsupported-operator = "ignore" # Ignore unsupported operators
invalid-type-form = "ignore" # Ignore invalid type formats
